Ratinal<64> and f64

+++ src/event/mod.rs
@@ -4,14 +4,14 @@ use pbr::ProgressBar;
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct Event {
     pub sounds: Vec<Sound>,
-    pub length: f32,
+    pub length: f64,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct Sound {
-    pub frequency: f32,
-    pub gain: f32,
-    pub pan: f32,
+    pub frequency: f64,
+    pub gain: f64,
+    pub pan: f64,
 }
 
 impl Sound {
@@ -25,7 +25,7 @@ impl Sound {
 }
 
 impl Event {
-    pub fn init(frequency: f32, gain: f32, pan: f32, length: f32) -> Event {
+    pub fn init(frequency: f64, gain: f64, pan: f64, length: f64) -> Event {
         Event {
             sounds: {
                 vec![Sound {
@@ -45,6 +45,7 @@ pub trait Render<T> {
 
 impl Render<Event> for Event {
     fn render(&mut self, oscillator: &mut Oscillator) -> StereoWaveform {
+//        println!("{:?}", self);
         oscillator.update(self.sounds.clone());
         let n_samples_to_generate = self.length * 44_100.0;
         oscillator.generate(n_samples_to_generate)
@@ -62,7 +63,7 @@ impl Render<Vec<Event>> for Vec<Event> {
                 pb.message("Rendering:  ");
                 let sub_div = 1 + n_events / 1000;
                 let mut i = 1;
-        for mut event in events {
+            for mut event in events {
                         i += 1;
                         if i % sub_div == 0 {
                             pb.add(sub_div as u64);
