+++ src/instrument/voice.rs
@@ -1,6 +1,6 @@
 use instrument::loudness::loudness_normalization;
-use std::f32::consts::PI;
-fn tau() -> f32 {
+use std::f64::consts::PI;
+fn tau() -> f64 {
     PI * 2.0
 }
 
@@ -9,13 +9,13 @@ pub struct Voice {
     pub index: usize,
     pub past: VoiceState,
     pub current: VoiceState,
-    pub phase: f32,
+    pub phase: f64,
 }
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct VoiceState {
-    pub frequency: f32,
-    pub gain: f32,
+    pub frequency: f64,
+    pub gain: f64,
 }
 
 impl VoiceState {
@@ -38,9 +38,9 @@ impl Voice {
     }
     pub fn generate_waveform(
         &mut self,
-        buffer: &mut Vec<f32>,
+        buffer: &mut Vec<f64>,
         portamento_length: usize,
-        factor: f32,
+        factor: f64,
     ) {
         let p_delta = self.calculate_portamento_delta(portamento_length);
         let g_delta = self.calculate_gain_delta(buffer.len());
@@ -52,7 +52,7 @@ impl Voice {
         }
     }
 
-    pub fn update(&mut self, mut frequency: f32, gain: f32) {
+    pub fn update(&mut self, mut frequency: f64, gain: f64) {
         if frequency < 20.0 {
             frequency = 0.0;
         }
@@ -78,32 +78,32 @@ impl Voice {
     pub fn generate_sample(
         &mut self,
         index: usize,
-        p_delta: f32,
-        g_delta: f32,
+        p_delta: f64,
+        g_delta: f64,
         portamento_length: usize,
-        factor: f32,
-    ) -> f32 {
+        factor: f64,
+    ) -> f64 {
         let frequency = if self.sound_to_silence() {
             self.past.frequency
         } else if index < portamento_length && !self.silence_to_sound() && !self.sound_to_silence()
         {
-            self.past.frequency + (index as f32 * p_delta)
+            self.past.frequency + (index as f64 * p_delta)
         } else {
             self.current.frequency
         };
 
-        let gain = (index as f32 * g_delta) + self.past.gain;
+        let gain = (index as f64 * g_delta) + self.past.gain;
         let current_phase = ((factor * frequency) + self.phase) % tau();
         self.phase = current_phase;
 
         current_phase.sin() * gain
     }
 
-    pub fn calculate_portamento_delta(&self, portamento_length: usize) -> f32 {
-        (self.current.frequency - self.past.frequency) / (portamento_length as f32)
+    pub fn calculate_portamento_delta(&self, portamento_length: usize) -> f64 {
+        (self.current.frequency - self.past.frequency) / (portamento_length as f64)
     }
 
-    pub fn calculate_gain_delta(&self, fade_length: usize) -> f32 {
-        (self.current.gain - self.past.gain) / (fade_length as f32)
+    pub fn calculate_gain_delta(&self, fade_length: usize) -> f64 {
+        (self.current.gain - self.past.gain) / (fade_length as f64)
     }
 }
