+++ src/settings.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Debug, PartialEq)]
 pub struct Settings {
-    pub sample_rate: f32,
+    pub sample_rate: f64,
     pub yin_buffer_size: usize,
     pub buffer_size: usize,
     pub probability_threshold: f32,
